TradeHistory rendering live data

+++ db/init.sql
@@ -51,9 +51,9 @@ CREATE TABLE IF NOT EXISTS transactions (
   id SERIAL PRIMARY KEY,
   order_type transaction_type,
   currency VARCHAR(10) NOT NULL,
-  purchase_price INTEGER NOT NULL,
-  total_trade_fiat INTEGER NOT NULL,
-  total_trade_coin INTEGER NOT NULL,
+  purchase_price DECIMAL NOT NULL,
+  total_trade_fiat DECIMAL NOT NULL,
+  total_trade_coin DECIMAL NOT NULL,
   order_datetime timestamp,
   trader_id INTEGER REFERENCES trader(id),
   coin_id INTEGER REFERENCES coins(id)
